Introducing KnpUIpsum
Head to your browser and go to http://localhost:8000. Say hello to The Space Bar! This is a Symfony application - the one we're building in our beginner Symfony series. Click into one of the articles to see a bunch of delightful, fake text that we're using to make this page look real. Each time you refresh, you get new random, happy content.

To find out where this is coming from, in the project, open src/Service/KnpUIpsum.php. Yes! This is our new creation: it returns 'lorem ipsum' dummy text, but with a little KnpUniversity flare: the classic latin is replaced with rainbows, unicorns, sunshine and more of our favorite things.

340 lines src/Service/KnpUIpsum.php
... lines 1 - 2
namespace App\Service;
... lines 4 - 9
class KnpUIpsum {
    private $ unicornsAreReal;
    private $ minSunshine;
    public function __construct(bool $ unicornsAreReal = true, $ minSunshine = 3)
    {
        $ this->unicornsAreReal = $ unicornsAreReal;
        $ this->minSunshine = $ minSunshine;
    }
    /**
     * Returns several paragraphs of random ipsum text.
     *
     * @param int $ count
     * @return string
     */
    public function getParagraphs(int  = 3): string{
        $ paragraphs = array();
        for ($ i = 0; $ i < ; $ i++) {
            $ paragraphs[] = $ this->addJoy(->getSentences($ this->gauss(5.8, 1.93)));
        }
        return implode(nn, $ paragraphs);
    }
... lines 37 - 338
}

And, you know what? I think we all deserve more cupcakes, kittens & baguettes in our life. So I want to share this functionality with the world, by creating the KnpULoremIpsumBundle! Yep, we're going to extract this class into its own bundle, handle configuration, add tests, and do a bunch of other cool stuff.

Right now, we're using this code inside of ArticleController: it's being passed to the constructor. Below, we use that to generate the content.

77 lines src/Controller/ArticleController.php
... lines 1 - 2
namespace App\Controller;
... lines 4 - 14
class ArticleController extends AbstractController{
    /**
     * Currently unused: just showing a controller with a constructor!
     */
    private $ isDebug;
    private $ knpUIpsum;
    public function __construct(bool $ isDebug, KnpUIpsum $ knpUIpsum)
    {
        $ this->isDebug = $ isDebug;
        $ this->knpUIpsum = $ knpUIpsum;
    }
    /**
     * @Route(/, name='app_homepage')
     */
    public function homepage()
    {
        return $ this->render('article/homepage.html.twig');
    }
... lines 37 - 75
}