Change the root key to use knp_u_lorem_ipsum instead as printed in the list of valid keys whem errors were displayed in previous commit. Next, open our extension class, var_dump($ configs) and die.
Try it out! No error! And cool! That bar: true value is passed to the load method! We're one step closer to using that config to tweak our service.

But, there are two weird things. First, the root key is... uh... not perfect. The knp_u_ is.. weird - I want it be knpu_... but apparently Symfony disagrees: our extra capital U is confusing things. We'll fix this in a bit.

The second weird thing is that the $ configs value that's passed to load() is not just a simple array with bar=true. Nope, it's an array of arrays. Inception. Why? Well, it's possible that the user could add configuration for our bundle in multiple files. Like, we could have a dev environment-specific YAML file. When that happens, instead of merging that config together, it would pass us the configuration from both files. For example, if knp_u_lorem_ipsum existed in three different files, this array would have three different arrays inside. And, yep! It will be our job to merge them together. But, that's actually going to be really cool.

But before we do that, let's fix our alias to be knpu_lorem_ipsum. It's not something you often need to worry about, but the fix is super interesting.