Creating the Configuration Class
Inside the DependencyInjection directory, create a new class called Configuration. Make this implement ConfigurationInterface: the one from the Config component. We'll need to implement one method: go to the Code -> Generate menu, or Cmd+N on a Mac, select 'Implement Methods' and choose getConfigBuilder().

This is one of the strangest classes you'll ever see. By using PHP code, we're going to define the entire tree of valid config that can be passed to our bundle.

A great way to see how this class works is to look at an existing one! Type Shift+Shift to open a class called FrameworkExtension, deep in the core of Symfony. Yep, this is the extension class for FrameworkBundle! It has the same load() method as our extension.

In the same directory, if you click on the top tree, you'll find a class called Configuration. Inside, it defines all of the valid config keys with a, sort of, nested tree format. This is a super powerful and, honestly, super complex system. We're only going to use a few basic features. If you need to define a more complex config tree, definitely steal, um, borrow, from these core classes.

Building the Config Tree

Back in our class, start with $ treeBuilder = new TreeBuilder(). Then, $ rootNode = $ treeBuilder->root() and pass the name of our key: knpu_lorem_ipsum. Now... just start building the config tree! $ rootNode->children(), and below, let's create two keys. The first will be for the 'unicorns are real' value, and it should be a boolean. To add that, say ->booleanNode('unicorns_are_real'), ->defaultTrue() and to finish configuring this node, ->end(). The other option will an integer: ->integerNode('min_sunshine'), default it to 3, then ->end(). Call ->end() one more time to finish the children(). Weird, right!? Return the $ treeBuilder at the bottom.

Using the Configuration Class

In our extension, we can use this to validate and merge all the config together. Start with $ configuration = $ this->getConfiguration() and pass this $ configs and the container. This simply instantiates the Configuration class. Here's the really important part: $ config = $ this->processConfiguration(): pass the configuration object and the original, raw array of $ configs. var_dump() that final config and die!

Let's see what happens! Find your browser and... refresh! We get an error... which is awesome! It says:

    Unrecognized option 'bar' under 'knpu_lorem_ipsum'

This is telling us:

    Yo! bar is not one of the valid config keys!