Making the Word Provider Configurable
So... we refactored our code to be more flexible... but it's still not possible for the user to override the word provider. Here's my idea: in the Configuration class, add a new scalar node - in other words, a string option - called word_provider. Default this to null, and you can add some documentation to be super cool. If the user wants to customize the word list, they will set this to the service id of their own word provider. So, in the extension class, if the that value is not set to null, let's replace the first argument entirely: $ definition->setArgument() with 0 and $ config['word_provider'].

We're not setting this config value yet, but when we refresh, great! We at least didn't break anything... though we do have a small mistake...