This is the perfect use-case for an interface! Wooo! In the bundle, create a new PHP class. Call it WordProviderInterface and change the 'kind' from class to interface.
Inside, add the getWordList() method and make it return an array. This is also the perfect place to add some documentation about what this method should do. With the interface done, go back to KnpUIpsum, change the type-hint to WordProviderInterface. The user can now pass anything they want, as long as it has this getWordList() method... because that is what we're using at the bottom of KnpUIpsum. Then, of course, we also need to go open our provider and make sure it implements this interface: implements WordProviderInterface.

If you try it now... not broken! And yea, our CustomUserProvider will still extend KnpUWordProvider, but that's now optional - we could just implement the interface directly.

Next, let's take a big step and move our bundle out of our code and give it it's own composer.json file!