Now, find your terminal, and run: 'composer update'
Perfect! Once that finishes, we can go back to our functional test. Re-type the 'l' on Kernel and... yes! There is the Kernel class from http-kernel.
This requires us to implement two methods. Go to the Code -> Generate menu - or Command + N on a Mac - click 'Implement Methods' and choose the two.
Inside registerBundles, return an array and only enable our bundle: new KnpULoremIpsumBundle(). Since we're not dependent on any other bundles - like FrameworkBundle - we should, in theory, be able to boot an app with only this. Kinda cool!
And... that's it! Our app is ready. Back in testServiceWiring, add $ kernel = new KnpULoremIpsumTestingKernel() and pass this test for the environment, thought that doesn't matter, and true for debug. Next, boot the kernel, and say $ container = $ kernel->getContainer().
This is great! We just booted a real Symfony app. And now, we can makes sure our service exists. Add $ ipsum = $ container->get(), copy the id of our service, and paste it here. We can do this because the service is public.

Let's add some very basic checks, like $ this->assertInstanceOf() that KnpUIpsum::class is the type of $ ipsum. And also, $ this->assertInternalType() that a string is what we get back when we call $ ipsum->getParagraphs().

The unit test truly tests this class - so we really only need a sanity check. I think it's time to try this! Find your terminal, and run:

./vendor/bin/simple-phpunit

Yes! We're now sure that our service is wired correctly! So, this functional test didn't fail like I promised in the last chapter. But the point is this: before we added our dependencies, our bundle was not actually setup correctly.

And, woh! In the tests/ directory, we suddenly have a cache/ folder! That comes from our kernel: it caches files just like a normal app. To make sure this doesn't get committed, open .gitignore and ignore /tests/cache.

Next, let's get a little more complex by testing that some of our configuration options work.