When you create a controller in a reusable bundle, the best practice is to register your controller as a proper service and use dependency injection to get anything you need.
Add public function __construct() and type-hint the first argument with KnpUIpsum. I'll press Alt+Enter and choose Initialize Fields so that PhpStorm creates and sets a property for that.
Down below, return $ this->json() - we will not have auto-complete for that method because of the missing AbstractController - with a paragraphs key set to $ this->knpUIpsum->getParagraphs() and a sentences key set to $ this->knpUIpsum->getSentences()
Excellent!

Registering your Controller as a Service

Next, we need to register this as a service. In services.xml, copy the first service, call this one ipsum_api_controller, and set its class name. For now, don't add public='true' or false: we'll learn more about this in a minute. Pass one argument: the main knpu_lorem_ipsum.knpu_ipsum service.
Perfect!